GeoCoding
This is the sample python program which is used for geocoding the addresses . User can pass the input file and specify the output file name
to store the geo output. This program uses the google API and user must have google API key in order to run it. This key needs to specify in the  API_KEY = '<Add Your Key Here>' in program.